+++ prisma/migrations/20231220073242_provide_default_value_0_for_likes_no/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "DefaultModel" ALTER COLUMN "likesNo" SET DEFAULT 0;
+
+-- AlterTable
+ALTER TABLE "VerificationEmail" ALTER COLUMN "expiration_time" SET DEFAULT NOW() + interval '5 minutes';