+++ prisma/seed.ts
@@ -28,10 +28,6 @@ const users = [
 
 const categories: { name: string; id: string }[] = [
     {
-        name: 'All things',
-        id: 'clpf9dsqc000008kwc9d3f8tr'
-    },
-    {
         name: 'Fashion',
         id: 'clpf9hn1t000108kw2iw00323'
     },
@@ -65,7 +61,6 @@ const generateModel = () => {
     return {
         id: faker.string.uuid(),
         categoryId: faker.helpers.arrayElement(categories).id,
-        likesNo: faker.number.int({ min: 0, max: 200 }),
         image: faker.image.url(),
         subImage1: faker.image.url(),
         subImage2: faker.image.url(),
@@ -148,7 +143,6 @@ async function handleDefaultModels() {
         return {
             model_id: defaultModel.id,
             category_id: defaultModel.categoryId,
-            likesNo: defaultModel.likesNo,
             imageUrl: defaultModel.image,
             subImageUrls: [defaultModel.subImage1, defaultModel.subImage2]
         };