+++ src/handlers/order.handler.ts
@@ -2,8 +2,9 @@ import { Handler } from '@interfaces';
 import { OrderListResultDto, OrderResultDto, UpdateOrderResultDto } from '@dtos/out';
 import { prisma } from '@repositories';
 import { USER_NOT_FOUND, ORDER_NOT_FOUND, UPDATE_ORDER_FAILED } from '@constants';
-import { UserRole } from '@prisma/client';
+import { Status, UserRole } from '@prisma/client';
 import { OrderQueryStringDto, UpdateOrderDto } from '@dtos/in';
+import { logger } from '@utils';
 
 const getOrderById: Handler<OrderResultDto, { Params: { orderId: string } }> = async (req, res) => {
     const userId = req.userId;
@@ -189,8 +190,105 @@ const update: Handler<UpdateOrderResultDto, { Params: { id: string }; Body: Upda
     }
 };
 
+//use this for payment with cash
+const pay: Handler<UpdateOrderResultDto, { Params: { id: string } }> = async (req, res) => {
+    try {
+        const order = await prisma.order.findUnique({
+            where: {
+                id: req.params.id
+            }
+        });
+
+        if (!order) {
+            return res.notFound('Order not found');
+        }
+
+        if (order.isPaid) {
+            return res.badRequest('Order is already paid');
+        }
+
+        const updatedOrder = await prisma.order.update({
+            data: {
+                isPaid: true
+            },
+            where: {
+                id: req.params.id
+            }
+        });
+
+        return {
+            creationTime: updatedOrder.creation_time.toISOString(),
+            digitalOrderId: updatedOrder.digital_order_id || undefined,
+            district: updatedOrder.district,
+            estimatedDeliveryTime: updatedOrder.est_deli_time.toISOString(),
+            id: req.params.id,
+            isPaid: updatedOrder.isPaid,
+            shippingFee: updatedOrder.shipping_fee,
+            status: updatedOrder.status,
+            street: updatedOrder.street,
+            streetNo: updatedOrder.streetNo,
+            totalPrice: updatedOrder.total_price,
+            ward: updatedOrder.ward,
+            note: updatedOrder.extra_note || undefined,
+            userId: updatedOrder.user_id
+        };
+    } catch (e) {
+        logger.error('Error when updating order paid:', e);
+        return res.internalServerError('Failed to updating order paid. Please try again later.');
+    }
+};
+
+const cancel: Handler<UpdateOrderResultDto, { Params: { id: string } }> = async (req, res) => {
+    try {
+        const order = await prisma.order.findUnique({
+            where: {
+                id: req.params.id
+            }
+        });
+
+        if (!order) {
+            return res.notFound('Order not found');
+        }
+
+        if (order.status !== Status.ORDER_PENDING) {
+            return res.badRequest('Order cannot be canceled as it is not in pending status');
+        }
+
+        const updatedOrder = await prisma.order.update({
+            data: {
+                status: Status.REJECTED
+            },
+            where: {
+                id: req.params.id
+            }
+        });
+
+        return {
+            creationTime: updatedOrder.creation_time.toISOString(),
+            digitalOrderId: updatedOrder.digital_order_id || undefined,
+            district: updatedOrder.district,
+            estimatedDeliveryTime: updatedOrder.est_deli_time.toISOString(),
+            id: req.params.id,
+            isPaid: updatedOrder.isPaid,
+            shippingFee: updatedOrder.shipping_fee,
+            status: updatedOrder.status,
+            street: updatedOrder.street,
+            streetNo: updatedOrder.streetNo,
+            totalPrice: updatedOrder.total_price,
+            ward: updatedOrder.ward,
+            note: updatedOrder.extra_note || undefined,
+            userId: updatedOrder.user_id
+        };
+    } catch (e) {
+        logger.error('Error when canceling order:', e);
+        return res.internalServerError('Failed to cancel the order. Please try again later.');
+    }
+};
+
 export const ordersHandler = {
     getOrderById,
     getOrders,
-    update
+    update,
+    pay,
+    cancel
 };